backend/requirements.txt
Flask-Cors==3.0.10
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
openpyxl==3.0.6
pandas==1.2.0
requests==2.22.0
SQLAlchemy==1.3.23
fhir.resources==6.1.0
fhir==0.0.4
itsdangerous==2.0.1
werkzeug==2.0
spacy==3.2.4
psycopg2-binary==2.9.3
